requirements.txt
# Python Version: 3.12.2
alembic==1.14.1
bcrypt==4.3.0
celery==5.4.0
fastapi==0.115.8
matplotlib==3.10.0
numpy==2.2.3
opencv-python-headless==4.11.0.86
passlib==1.7.4
pika==1.3.2
pillow==11.1.0
plotly==6.0.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
scikit-learn==1.6.1
SQLAlchemy==2.0.38
tifffile==2025.2.18
uvicorn==0.34.0
jupyter
pytest
httpx
pytest-mock
pytest-asyncio

